Module mozjs_sys::generated::root::js::gc

Modules

detail

Structs

Arena

Cell

ChunkBase

GCRuntime

MarkBitmap

PersistentRootedMarker

StoreBuffer

TenuredCell

TenuredChunk

TenuredChunkBase

TenuredChunkInfo

Enums

ColorBit

MarkColor

PerformanceHint

Constants

MarkBitmap_WordCount

Functions

FinalizeDeadNurseryObject⚠

NewMemoryInfoObject⚠

Create an object providing access to the garbage collector’s internal notion of the current state of memory (both GC heap memory and GCthing-controlled malloc memory.

PerformIncrementalReadBarrier⚠

RealmNeedsSweep⚠

SetPerformanceHint⚠

TraceExternalEdge⚠

TraceExternalEdge1⚠

TraceExternalEdge2⚠

TraceExternalEdge3⚠

TraceExternalEdge4⚠

TraceExternalEdge5⚠

TraceExternalEdge6⚠

TraceExternalEdge7⚠

TraceExternalEdge8⚠

TraceExternalEdge9⚠

TraceRealm⚠

Type Definitions

DecommitBitmap

MarkBitmapWord