mozjs_sys::jsapi::JS

Struct TransitiveCompileOptions

#[repr(C, packed(1))]
pub struct TransitiveCompileOptions {

    pub filename_: ConstUTF8CharsZ,
    pub introducerFilename_: ConstUTF8CharsZ,
    pub sourceMapURL_: *const u16,
    pub mutedErrors_: bool,
    pub forceStrictMode_: bool,
    pub alwaysUseFdlibm_: bool,
    pub skipFilenameValidation_: bool,
    pub hideScriptFromDebugger_: bool,
    pub deferDebugMetadata_: bool,
    pub eagerDelazificationStrategy_: DelazificationOption,
    pub eagerBaselineStrategy_: EagerBaselineOption,
    pub selfHostingMode: bool,
    pub discardSource: bool,
    pub sourceIsLazy: bool,
    pub allowHTMLComments: bool,
    pub nonSyntacticScope: bool,
    pub topLevelAwait: bool,
    pub borrowBuffer: bool,
    pub usePinnedBytecode: bool,
    pub prefableOptions_: PrefableCompileOptions,
    pub introductionType: *const c_char,
    pub introductionLineno: c_uint,
    pub introductionOffset: u32,
    pub hasIntroductionInfo: bool,

}

The common base class for the CompileOptions hierarchy.

Use this in code that needs to propagate compile options from one compilation unit to another.

Fields
filename_: ConstUTF8CharsZintroducerFilename_: ConstUTF8CharsZsourceMapURL_: *const u16mutedErrors_: bool

The Web Platform allows scripts to be loaded from arbitrary cross-origin sources. This allows an attack by which a malicious website loads a sensitive file (say, a bank statement) cross-origin (using the user’s cookies), and sniffs the generated syntax errors (via a window.onerror handler) for juicy morsels of its contents.

To counter this attack, HTML5 specifies that script errors should be sanitized (“muted”) when the script is not same-origin with the global for which it is loaded. Callers should set this flag for cross-origin scripts, and it will be propagated appropriately to child scripts and passed back in JSErrorReports.

forceStrictMode_: boolalwaysUseFdlibm_: boolskipFilenameValidation_: boolhideScriptFromDebugger_: booldeferDebugMetadata_: booleagerDelazificationStrategy_: DelazificationOptioneagerBaselineStrategy_: EagerBaselineOptionselfHostingMode: booldiscardSource: boolsourceIsLazy: boolallowHTMLComments: boolnonSyntacticScope: booltopLevelAwait: boolborrowBuffer: boolusePinnedBytecode: boolprefableOptions_: PrefableCompileOptionsintroductionType: *const c_char

|introductionType| is a statically allocated C string. See JSScript.h for more information.

introductionLineno: c_uintintroductionOffset: u32hasIntroductionInfo: bool