Type Definition mozjs_sys::jsapi::JS::DispatchToEventLoopCallback

type DispatchToEventLoopCallback = Option<unsafe extern "C" fn(closure: *mut c_void, dispatchable: *mut Dispatchable) -> bool>;

Callback to dispatch a JS::Dispatchable to a JSContext's thread's event loop.

The DispatchToEventLoopCallback set on a particular JSContext must accept JS::Dispatchable instances and arrange for their run methods to be called eventually on the JSContext's thread. This is used for cross-thread dispatch, so the callback itself must be safe to call from any thread.

If the callback returns true, it must eventually run the given Dispatchable; otherwise, SpiderMonkey may leak memory or hang.

The callback may return false to indicate that the JSContext's thread is shutting down and is no longer accepting runnables. Shutting down is a one-way transition: once the callback has rejected a runnable, it must reject all subsequently submitted runnables as well.

To establish a DispatchToEventLoopCallback, the embedding may either call InitDispatchToEventLoop to provide its own, or call js::UseInternalJobQueues to select a default implementation built into SpiderMonkey. This latter depends on the embedding to call js::RunJobs on the JavaScript thread to process queued Dispatchables at appropriate times.