Enum mozjs_sys::jsapi::JSGCParamKey

#[repr(u32)]pub enum JSGCParamKey {
    JSGC_MAX_BYTES,
    JSGC_MAX_NURSERY_BYTES,
    JSGC_BYTES,
    JSGC_NUMBER,
    JSGC_MODE,
    JSGC_UNUSED_CHUNKS,
    JSGC_TOTAL_CHUNKS,
    JSGC_SLICE_TIME_BUDGET_MS,
    JSGC_MARK_STACK_LIMIT,
    JSGC_HIGH_FREQUENCY_TIME_LIMIT,
    JSGC_HIGH_FREQUENCY_LOW_LIMIT,
    JSGC_HIGH_FREQUENCY_HIGH_LIMIT,
    JSGC_HIGH_FREQUENCY_HEAP_GROWTH_MAX,
    JSGC_HIGH_FREQUENCY_HEAP_GROWTH_MIN,
    JSGC_LOW_FREQUENCY_HEAP_GROWTH,
    JSGC_DYNAMIC_HEAP_GROWTH,
    JSGC_DYNAMIC_MARK_SLICE,
    JSGC_ALLOCATION_THRESHOLD,
    JSGC_MIN_EMPTY_CHUNK_COUNT,
    JSGC_MAX_EMPTY_CHUNK_COUNT,
    JSGC_COMPACTING_ENABLED,
    JSGC_NON_INCREMENTAL_FACTOR,
    JSGC_AVOID_INTERRUPT_FACTOR,
    JSGC_NURSERY_FREE_THRESHOLD_FOR_IDLE_COLLECTION,
    JSGC_PRETENURE_THRESHOLD,
    JSGC_PRETENURE_GROUP_THRESHOLD,
    JSGC_NURSERY_FREE_THRESHOLD_FOR_IDLE_COLLECTION_PERCENT,
    JSGC_MIN_NURSERY_BYTES,
    JSGC_MIN_LAST_DITCH_GC_PERIOD,
    JSGC_ZONE_ALLOC_DELAY_KB,
    JSGC_NURSERY_BYTES,
    JSGC_MALLOC_THRESHOLD_BASE,
    JSGC_MALLOC_GROWTH_FACTOR,
}

Variants

JSGC_MAX_BYTES

Maximum nominal heap before last ditch GC.

Soft limit on the number of bytes we are allowed to allocate in the GC heap. Attempts to allocate gcthings over this limit will return null and subsequently invoke the standard OOM machinery, independent of available physical memory.

Pref: javascript.options.mem.max Default: 0xffffffff

JSGC_MAX_NURSERY_BYTES

Maximum size of the generational GC nurseries.

This will be rounded to the nearest gc::ChunkSize.

Pref: javascript.options.mem.nursery.max_kb Default: JS::DefaultNurseryMaxBytes

JSGC_BYTES

Amount of bytes allocated by the GC.

JSGC_NUMBER

Number of times GC has been invoked. Includes both major and minor GC.

JSGC_MODE

Select GC mode.

See: JSGCMode in GCAPI.h prefs: javascript.options.mem.gc_per_zone and javascript.options.mem.gc_incremental. Default: JSGC_MODE_ZONE_INCREMENTAL

JSGC_UNUSED_CHUNKS

Number of cached empty GC chunks.

JSGC_TOTAL_CHUNKS

Total number of allocated GC chunks.

JSGC_SLICE_TIME_BUDGET_MS

Max milliseconds to spend in an incremental GC slice.

Pref: javascript.options.mem.gc_incremental_slice_ms Default: DefaultTimeBudgetMS.

JSGC_MARK_STACK_LIMIT

Maximum size the GC mark stack can grow to.

Pref: none Default: MarkStack::DefaultCapacity

JSGC_HIGH_FREQUENCY_TIME_LIMIT

GCs less than this far apart in milliseconds will be considered 'high-frequency GCs'.

Pref: javascript.options.mem.gc_high_frequency_time_limit_ms Default: HighFrequencyThreshold

JSGC_HIGH_FREQUENCY_LOW_LIMIT

Start of dynamic heap growth (MB).

Pref: javascript.options.mem.gc_high_frequency_low_limit_mb Default: HighFrequencyLowLimitBytes

JSGC_HIGH_FREQUENCY_HIGH_LIMIT

End of dynamic heap growth (MB).

Pref: javascript.options.mem.gc_high_frequency_high_limit_mb Default: HighFrequencyHighLimitBytes

JSGC_HIGH_FREQUENCY_HEAP_GROWTH_MAX

Upper bound of heap growth percentage.

Pref: javascript.options.mem.gc_high_frequency_heap_growth_max Default: HighFrequencyHeapGrowthMax

JSGC_HIGH_FREQUENCY_HEAP_GROWTH_MIN

Lower bound of heap growth percentage.

Pref: javascript.options.mem.gc_high_frequency_heap_growth_min Default: HighFrequencyHeapGrowthMin

JSGC_LOW_FREQUENCY_HEAP_GROWTH

Heap growth percentage for low frequency GCs.

Pref: javascript.options.mem.gc_low_frequency_heap_growth Default: LowFrequencyHeapGrowth

JSGC_DYNAMIC_HEAP_GROWTH

If false, the heap growth factor is fixed at 3. If true, it is determined based on whether GCs are high- or low- frequency.

Pref: javascript.options.mem.gc_dynamic_heap_growth Default: DynamicHeapGrowthEnabled

JSGC_DYNAMIC_MARK_SLICE

If true, high-frequency GCs will use a longer mark slice.

Pref: javascript.options.mem.gc_dynamic_mark_slice Default: DynamicMarkSliceEnabled

JSGC_ALLOCATION_THRESHOLD

Lower limit for collecting a zone.

Zones smaller than this size will not normally be collected.

Pref: javascript.options.mem.gc_allocation_threshold_mb Default GCZoneAllocThresholdBase

JSGC_MIN_EMPTY_CHUNK_COUNT

We try to keep at least this many unused chunks in the free chunk pool at all times, even after a shrinking GC.

Pref: javascript.options.mem.gc_min_empty_chunk_count Default: MinEmptyChunkCount

JSGC_MAX_EMPTY_CHUNK_COUNT

We never keep more than this many unused chunks in the free chunk pool.

Pref: javascript.options.mem.gc_min_empty_chunk_count Default: MinEmptyChunkCount

JSGC_COMPACTING_ENABLED

Whether compacting GC is enabled.

Pref: javascript.options.mem.gc_compacting Default: CompactingEnabled

JSGC_NON_INCREMENTAL_FACTOR

Percentage for how far over a trigger threshold we go before triggering a non-incremental GC.

We trigger an incremental GC when a trigger threshold is reached but the collection may not be fast enough to keep up with the mutator. At some point we finish the collection non-incrementally.

Default: NonIncrementalFactor Pref: None

JSGC_AVOID_INTERRUPT_FACTOR

Percentage for how far over a trigger threshold we go before triggering an incremental collection that would reset an in-progress collection.

Default: AvoidInterruptFactor Pref: None

JSGC_NURSERY_FREE_THRESHOLD_FOR_IDLE_COLLECTION

Attempt to run a minor GC in the idle time if the free space falls below this number of bytes.

Default: NurseryChunkUsableSize / 4 Pref: None

JSGC_PRETENURE_THRESHOLD

If this percentage of the nursery is tenured and the nursery is at least 4MB, then proceed to examine which groups we should pretenure.

Default: PretenureThreshold Pref: None

JSGC_PRETENURE_GROUP_THRESHOLD

If the above condition is met, then any object group that tenures more than this number of objects will be pretenured (if it can be).

Default: PretenureGroupThreshold Pref: None

JSGC_NURSERY_FREE_THRESHOLD_FOR_IDLE_COLLECTION_PERCENT

Attempt to run a minor GC in the idle time if the free space falls below this percentage (from 0 to 99).

Default: 25 Pref: None

JSGC_MIN_NURSERY_BYTES

Minimum size of the generational GC nurseries.

This value will be rounded to the nearest Nursery::SubChunkStep if below gc::ChunkSize, otherwise it'll be rounded to the nearest gc::ChunkSize.

Default: Nursery::SubChunkLimit Pref: javascript.options.mem.nursery.min_kb

JSGC_MIN_LAST_DITCH_GC_PERIOD

The minimum time to allow between triggering last ditch GCs in seconds.

Default: 60 seconds Pref: None

JSGC_ZONE_ALLOC_DELAY_KB

The delay (in heapsize kilobytes) between slices of an incremental GC.

Default: ZoneAllocDelayBytes

JSGC_NURSERY_BYTES

The delay (in heapsize kilobytes) between slices of an incremental GC.

Default: ZoneAllocDelayBytes

JSGC_MALLOC_THRESHOLD_BASE

Retained size base value for calculating malloc heap threshold.

Default: MallocThresholdBase

JSGC_MALLOC_GROWTH_FACTOR

Growth factor for calculating malloc heap threshold.

Default: MallocGrowthFactor