Function mozjs_sys::jsapi::JS_EncodeStringToUTF8BufferPartial

pub unsafe extern "C" fn JS_EncodeStringToUTF8BufferPartial(
    cx: *mut JSContext, 
    str_: *mut JSString, 
    buffer: [u64; 2]
) -> u8

Encode as many scalar values of the string as UTF-8 as can fit into the caller-provided buffer replacing unpaired surrogates with the REPLACEMENT CHARACTER.

If JS::StringHasLatin1Chars(str) returns true, the function is guaranteed to convert the entire string if buffer.Length() >= 2 * JS_GetStringLength(str). Otherwise, the function is guaranteed to convert the entire string if buffer.Length() >= 3 * JS_GetStringLength(str).

This function does not alter the representation of |str| or any |JSString*| substring that is a constituent part of it. Returns mozilla::Nothing() on OOM, without reporting an error; some data may have been written to |buffer| when this happens.

If there’s no OOM, returns the number of code units read and the number of code units written.

The semantics of this method match the semantics of TextEncoder.encodeInto().

The function does not store an additional zero byte.