mozjs_sys::jsapi::js

Module gc

Modules

detail

Structs

Arena

ArenaChunk

ArenaChunkBase

ArenaChunkInfo

AtomicBitmap_Iter

Cell

ChunkBase

ChunkMarkBitmap

GCRuntime

ProfilerMemoryCounts

SharedMemoryUse

StoreBuffer

TenuredCell

Enums

ChunkKind

ColorBit

MarkColor

Constants

AtomicBitmap_BitsPerWord

ChunkMarkBitmap_FirstThingAdjustmentBits

ChunkMarkBitmap_FirstThingAdjustmentWords

Statics

AtomicBitmap_BitCount⚠

AtomicBitmap_WordCount⚠

Functions

GetGCContext⚠

GetProfilerMemoryCounts⚠

NewMemoryInfoObject⚠

Create an object providing access to the garbage collector’s internal notion of the current state of memory (both GC heap memory and GCthing-controlled malloc memory.

PerformIncrementalReadBarrier⚠

TraceExternalEdge⚠

TraceExternalEdge1⚠

TraceExternalEdge2⚠

TraceExternalEdge3⚠

TraceExternalEdge4⚠

TraceExternalEdge5⚠

TraceExternalEdge6⚠

TraceExternalEdge7⚠

TraceExternalEdge8⚠

TraceExternalEdge9⚠

TraceRealmRoot⚠

Type Aliases

AtomicBitmapWord

Atomic implementation for integral types.

AtomicBitmap_Word

Atomic implementation for integral types.

ChunkArenaBitmap

An object like std::bitset but which provides access to the underlying storage.

ChunkMarkBitmap_Bitmap

ChunkPageBitmap

An object like std::bitset but which provides access to the underlying storage.

SharedMemoryMap