Module mozjs_sys::jsapi::js::shadow

Structs

BaseShape

Function

Object
This layout is shared by all native objects. For non-native objects, the group may always be accessed safely, and other members may be as well, depending on the object's specific layout.

ObjectGroup

Shape

Constants

Object_MAX_FIXED_SLOTS

Shape_FIXED_SLOTS_MASK

Shape_FIXED_SLOTS_SHIFT