mozjs_sys::jsapi

Module mozilla

Modules

baseprofiler

detail

external

ipc

literals

profiler

span_details

tl

Structs

AlignedStorage2

AlignmentFinder

AutoCleanLinkedList

BaseMarkerType

BaseTimeDuration

Instances of this class represent the length of an interval of time. Negative durations are allowed, meaning the end is before the start.

BaseTimeDurationPlatformUtils

Platform-specific implementation details of BaseTimeDuration.

BaseTimeDuration__SomethingVeryRandomHere

BitSet_Reference

BufferList

BufferList_IterImpl

BufferList_Segment

CStringHasher

CheckedInt

@class CheckedInt @brief Integer wrapper class checking for integer overflow and other errors @param T the integer type to wrap. Can be any type among the following: - any basic integer type such as |int| - any stdint type such as |int8_t|

CompactPair

CompactPair is the logical concatenation of an instance of A with an instance B. Space is conserved when possible. Neither A nor B may be a final class.

ContiguousEnumSize

ContiguousEnumValues

CorruptionCanary

This class is designed to cause crashes when various kinds of memory corruption are observed. For instance, let’s say we have a class C where we suspect out-of-bounds writes to some members. We can insert a member of type Poison near the members we suspect are being corrupted by out-of-bounds writes. Or perhaps we have a class K we suspect is subject to use-after-free violations, in which case it doesn’t particularly matter where in the class we add the member of type Poison.

CorruptionCanaryForStatics

A version of CorruptionCanary that is suitable as a member of objects that are statically allocated.

DebugOnly

DebugOnly contains a value of type T, but only in debug builds. In release builds, it does not contain a value. This helper is intended to be used with MOZ_ASSERT()-style macros, allowing one to write:

DefaultDelete

A default deletion policy using plain old operator delete.

DefaultHasher

EnumSet

EnumSet<T, U> is a set of values defined by an enumeration. It is implemented using a bit mask with the size of U for each value. It works both for enum and enum class types. EnumSet also works with U being a BitSet.

EnumSet_ConstIterator

EnumTypeFitsWithin

ErrorPropagationTag

A tag used to differentiate between GenericErrorResult created by the Err function (completely new error) and GenericErrorResult created by the Result::propagateErr function (propagated error). This can be used to track error propagation and eventually produce error stacks for logging/debugging purposes.

FailureLatch

FailureLatchInfallibleSource

FailureLatchSource

FailureLatch__bindgen_vtable

FallibleHashMethods

FloatingPoint

GenericErrorResult

A type that auto-converts to an error Result. This is like a Result without a success type. It’s the best return type for functions that always return an error–functions designed to build and populate error objects. It’s also useful in error-handling macros; see MOZ_TRY for an example.

HashCodeScrambler

A pseudorandom function mapping 32-bit integers to 32-bit integers.

HashCodeScrambler_SipHasher

HashMap

HashMapEntry

HashMap_MapHashPolicy

HashSet

HashSet_SetHashPolicy

JSONWriteFunc

JSONWriteFunc__bindgen_vtable

JSONWriter

JSONWriter_EscapedString

LinkedList

LinkedListElement

LinkedList_Iterator

MallocAllocPolicy

MarkerCategory

MarkerInnerWindowId

MarkerOptions

MarkerSchema

MarkerSchema_DynamicData

MarkerSchema_GraphData

MarkerSchema_PayloadField

MarkerSchema_SpecialFrontendLocation

MarkerSchema_StaticData

MarkerStack

MarkerThreadId

MarkerTiming

MaxContiguousEnumValue

MaxEnumValue

Maybe

Maybe_SomeGuard

MinContiguousEnumValue

MovingNotNull

NeverAllocPolicy

NotNull

Nothing

Ok

Empty struct, indicating success for operations that have no return value. For example, if you declare another empty struct struct OutOfMemory {};, then Result<Ok, OutOfMemory> represents either success or OOM.

Opaque

Opaque is a replacement for integral T in cases where only comparisons must be supported, and it’s desirable to prevent accidental dependency on exact values.

OwningNonNull

PointerHasher

PositionOfSignBit

PrintfTarget

PrintfTarget__bindgen_vtable

ProfileBufferBlockIndex

ProfileBufferChunk

ProfileBufferChunkManager

ProfileBufferChunkManagerSingle

ProfileBufferChunkManager__bindgen_vtable

ProfileBufferChunk_Header

ProfileBufferChunk_InternalHeader

ProfileBufferChunk_ReserveReturn

ProfileBufferEntryReader

ProfileBufferEntryReader_Deserializer

ProfileBufferEntryReader_DoubleSpanOfConstBytes

ProfileBufferEntryWriter

ProfileBufferEntryWriter_Serializer

ProfileBufferRawPointer

ProfileBufferUnownedCString

ProfileChunkedBuffer

ProfileChunkedBuffer_BlockIterator

ProfileChunkedBuffer_Reader

ProfileChunkedBuffer_RequestedChunkRefCountedHolder

ProfileChunkedBuffer_State

ProgressLogger

ProgressLogger_IndexAndProgressLoggerRange

ProgressLogger_IndexAndProgressLoggerRange_IndexAndProgressLogger

ProgressLogger_IndexAndProgressLoggerRange_IndexAndProgressLoggerEndIterator

ProgressLogger_IndexAndProgressLoggerRange_IndexAndProgressLoggerIterator

ProgressLogger_SharedProgress

ProportionValue

ProportionValue_Internal

Range

RangedPtr

ReentrancyGuard

RefCounted

RefPtrTraits

Result

Result<V, E> represents the outcome of an operation that can either succeed or fail. It contains either a success value of type V or an error value of type E.

ScopeExit

Span_storage_type

SprintfState

StaticLocalRefPtr

StaticRefPtr

StringBuffer

This structure precedes the string buffers “we” allocate. It may be the case that nsTAString::mData does not point to one of these special buffers. The mDataFlags member variable distinguishes the buffer type.

TimeDurationValueCalculator

Perform arithmetic operations on the value of a BaseTimeDuration without doing strict checks on the range of values.

TimeStamp

TimeStampTests

ULEB128Reader

UniquePtr

UniquePtr is a smart pointer that wholly owns a resource. Ownership may be transferred out of a UniquePtr through explicit action, but otherwise the resource is destroyed when the UniquePtr is destroyed.

Variant

mozilla::Variant

VariantType

Vector_CapacityAndReserved

Vector_ConstRange

Vector_Range

unused_t

Enums

JSONWriter_CollectionStyle

LinkedListElement_NodeKind

MarkerSchema_ETWMarkerGroup

MarkerSchema_Format

MarkerSchema_GraphColor

MarkerSchema_GraphType

MarkerSchema_InputType

MarkerSchema_Location

MarkerSchema_PayloadFlags

MarkerSchema_Searchable

MarkerTiming_Phase

MemoryOrdering

An enum of memory ordering possibilities for atomics.

NotNullTag

ProfileChunkedBuffer_RequestedChunkRefCountedHolder_State

ProfileChunkedBuffer_ThreadSafety

ProfilerStringView_Ownership

StackCaptureOptions

Vector__bindgen_ty_1

Constants

CorruptionCanaryForStatics_kCanarySet

MarkerInnerWindowId_scNoId

ProfileBufferChunkManager_scExpectedMaximumStackSize

ProportionValue_scFractionalBits

ProportionValue_scInvMaxD

ProportionValue_scInvalidU

ProportionValue_scMaxD

ProportionValue_scMaxU

Vector_InlineLength

Statics

JSONWriter_scArrayBeginString⚠

JSONWriter_scArrayEndString⚠

JSONWriter_scCommaString⚠

JSONWriter_scEmptyString⚠

JSONWriter_scFalseString⚠

JSONWriter_scNewLineString⚠

JSONWriter_scNullString⚠

JSONWriter_scObjectBeginString⚠

JSONWriter_scObjectEndString⚠

JSONWriter_scPropertyBeginString⚠

JSONWriter_scPropertyEndString⚠

JSONWriter_scQuoteString⚠

JSONWriter_scSpaceString⚠

JSONWriter_scTopObjectBeginString⚠

JSONWriter_scTopObjectEndString⚠

JSONWriter_scTrueString⚠

ProgressLogger_NO_LOCATION_UPDATE⚠

ProgressLogger_SharedProgress_NO_LOCATION_UPDATE⚠

VariantIndex_index⚠

Type Aliases

Array_ElementType

Array_const_iterator

Array_const_reverse_iterator

Array_iterator

Array_reverse_iterator

AutoCleanLinkedList_ClientType

AutoCleanLinkedList_Traits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

BitSet_Word

CStringHasher_Key

CStringHasher_Lookup

CheckedInt8

CheckedInt16

CheckedInt32

CheckedInt64

CheckedUint8

CheckedUint16

CheckedUint32

@class CheckedInt @brief Integer wrapper class checking for integer overflow and other errors @param T the integer type to wrap. Can be any type among the following: - any basic integer type such as |int| - any stdint type such as |int8_t|

CheckedUint64

CompactPair_Base

DefaultHasher_Lookup

EnumSet_serializedType

EnumSet_valueType

EnumeratedArray_ArrayType

EnumeratedArray_const_iterator

EnumeratedArray_const_reverse_iterator

EnumeratedArray_iterator

EnumeratedArray_reverse_iterator

FloatingPoint_Base

FloatingPoint_Bits

An unsigned integral type suitable for accessing the bitwise representation of T.

Generation

Opaque is a replacement for integral T in cases where only comparisons must be supported, and it’s desirable to prevent accidental dependency on exact values.

HashMapEntry_KeyType

HashMapEntry_ValueType

HashMap_AddPtr

HashMap_Entry

HashMap_Enum

HashMap_Impl

HashMap_Iterator

HashMap_Lookup

HashMap_MapHashPolicy_Base

HashMap_MapHashPolicy_KeyType

HashMap_ModIterator

HashMap_Ptr

HashMap_Range

HashMap_TableEntry

HashNumber

HashSet_AddPtr

HashSet_Entry

HashSet_Enum

HashSet_Impl

HashSet_Iterator

HashSet_Lookup

HashSet_ModIterator

HashSet_Ptr

HashSet_Range

HashSet_SetHashPolicy_Base

HashSet_SetHashPolicy_KeyType

InfinityBits_Traits

LinkedListElement_ClientType

LinkedListElement_ConstClientType

LinkedListElement_ConstRawType

LinkedListElement_RawType

LinkedListElement_Traits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

LinkedList_ClientType

LinkedList_ConstClientType

LinkedList_ConstElementType

LinkedList_ConstRawType

LinkedList_ElementType

LinkedList_Iterator_difference_type

LinkedList_Iterator_iterator_category

LinkedList_Iterator_pointer

LinkedList_Iterator_reference

LinkedList_Iterator_value_type

LinkedList_RawType

LinkedList_Traits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

LinkedList_const_iterator

LinkedList_iterator

MallocSizeOf

MarkerSchema_DataRow

mozilla::Variant

MarkerSchema_DataRowVector

Maybe_ValueType

PointerHasher_Lookup

ProfileBufferChunkManagerSingle_Length

ProfileBufferChunk_Byte

ProfileBufferChunk_Length

ProfileBufferChunk_SpanOfBytes

Span - slices for C++

ProfileBufferEntryReader_Byte

ProfileBufferEntryReader_Length

ProfileBufferEntryReader_SpanOfConstBytes

Span - slices for C++

ProfileBufferEntryReader_difference_type

ProfileBufferEntryReader_iterator_category

ProfileBufferEntryReader_pointer

ProfileBufferEntryReader_reference

ProfileBufferEntryReader_value_type

ProfileBufferEntryWriter_Byte

ProfileBufferEntryWriter_Length

ProfileBufferEntryWriter_SpanOfBytes

Span - slices for C++

ProfileBufferEntryWriter_iterator_category

ProfileBufferEntryWriter_pointer

ProfileBufferEntryWriter_reference

ProfileBufferEntryWriter_value_type

ProfileBufferIndex

ProfileChunkedBuffer_Byte

ProfileChunkedBuffer_Length

ProportionValue_UnderlyingType

Result_Impl

Result_err_type

Result_ok_type

SmprintfPointer

SmprintfPolicyPointer

Span_const_iterator

Span_const_reverse_iterator

Span_element_type

Span_index_type

Span_iterator

Span_pointer

Span_reference

Span_reverse_iterator

Span_value_type

SpecificNaNBits_Traits

TimeDuration

Specialization of BaseTimeDuration that uses TimeDurationValueCalculator for arithmetic on the mValue member.

TimeStampValue

TimeStamp_DurationType

Instances of this class represent the length of an interval of time. Negative durations are allowed, meaning the end is before the start.

UniqueFreePtr

UniquePtr_DeleterType

UniquePtr_ElementType

UniquePtr_Pointer

VariantType_Type

Variant_Impl

Variant_Tag

Vector_ElementType

Vector_Impl

fallible_t

Unions

AlignedStorage2_U

Utf8Unit

A code unit within a UTF-8 encoded string. (A code unit is the smallest unit within the Unicode encoding of a string. For UTF-8 this is an 8-bit number; for UTF-16 it would be a 16-bit number.)