mozjs_sys::jsapi::mozilla

Type Alias AutoCleanLinkedList_Traits

pub type AutoCleanLinkedList_Traits = LinkedListElementTraits;

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

Aliased Type

#[repr(C)]
pub struct AutoCleanLinkedList_Traits {
    pub _address: u8,
}

Fields
_address: u8