mozjs_sys::jsapi

Struct JSJitInfo

#[repr(C)]
pub struct JSJitInfo {
    pub __bindgen_anon_1: JSJitInfo__bindgen_ty_1,
    pub __bindgen_anon_2: JSJitInfo__bindgen_ty_2,
    pub __bindgen_anon_3: JSJitInfo__bindgen_ty_3,
    pub _bitfield_align_1: [u16; 0],
    pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4]>,
}

This struct contains metadata passed from the DOM to the JS Engine for JIT optimizations on DOM property accessors.

Eventually, this should be made available to general JSAPI users as not experimental and not a friend API, but weโre not ready to do so yet.

Fields
__bindgen_anon_1: JSJitInfo__bindgen_ty_1__bindgen_anon_2: JSJitInfo__bindgen_ty_2__bindgen_anon_3: JSJitInfo__bindgen_ty_3_bitfield_align_1: [u16; 0]_bitfield_1: __BindgenBitfieldUnit<[u8; 4]>
Implementations

impl JSJitInfo

pub fn type_(&self) -> u32

pub fn set_type_(&mut self, val: u32)

pub unsafe fn type__raw(this: *const Self) -> u32

pub unsafe fn set_type__raw(this: *mut Self, val: u32)

pub fn aliasSet_(&self) -> u32

pub fn set_aliasSet_(&mut self, val: u32)

pub unsafe fn aliasSet__raw(this: *const Self) -> u32

pub unsafe fn set_aliasSet__raw(this: *mut Self, val: u32)

pub fn returnType_(&self) -> u32

pub fn set_returnType_(&mut self, val: u32)

pub unsafe fn returnType__raw(this: *const Self) -> u32

pub unsafe fn set_returnType__raw(this: *mut Self, val: u32)

pub fn isInfallible(&self) -> u32

pub fn set_isInfallible(&mut self, val: u32)

pub unsafe fn isInfallible_raw(this: *const Self) -> u32

pub unsafe fn set_isInfallible_raw(this: *mut Self, val: u32)

pub fn isMovable(&self) -> u32

pub fn set_isMovable(&mut self, val: u32)

pub unsafe fn isMovable_raw(this: *const Self) -> u32

pub unsafe fn set_isMovable_raw(this: *mut Self, val: u32)

pub fn isEliminatable(&self) -> u32

pub fn set_isEliminatable(&mut self, val: u32)

pub unsafe fn isEliminatable_raw(this: *const Self) -> u32

pub unsafe fn set_isEliminatable_raw(this: *mut Self, val: u32)

pub fn isAlwaysInSlot(&self) -> u32

pub fn set_isAlwaysInSlot(&mut self, val: u32)

pub unsafe fn isAlwaysInSlot_raw(this: *const Self) -> u32

pub unsafe fn set_isAlwaysInSlot_raw(this: *mut Self, val: u32)

pub fn isLazilyCachedInSlot(&self) -> u32

pub fn set_isLazilyCachedInSlot(&mut self, val: u32)

pub unsafe fn isLazilyCachedInSlot_raw(this: *const Self) -> u32

pub unsafe fn set_isLazilyCachedInSlot_raw(this: *mut Self, val: u32)

pub fn isTypedMethod(&self) -> u32

pub fn set_isTypedMethod(&mut self, val: u32)

pub unsafe fn isTypedMethod_raw(this: *const Self) -> u32

pub unsafe fn set_isTypedMethod_raw(this: *mut Self, val: u32)

pub fn slotIndex(&self) -> u32

pub fn set_slotIndex(&mut self, val: u32)

pub unsafe fn slotIndex_raw(this: *const Self) -> u32

pub unsafe fn set_slotIndex_raw(this: *mut Self, val: u32)

pub fn new_bitfield_1( type_: u32, aliasSet_: u32, returnType_: u32, isInfallible: u32, isMovable: u32, isEliminatable: u32, isAlwaysInSlot: u32, isLazilyCachedInSlot: u32, isTypedMethod: u32, slotIndex: u32, ) -> __BindgenBitfieldUnit<[u8; 4]>