Struct mozjs_sys::jsapi::JSStructuredCloneData

#[repr(C)]
pub struct JSStructuredCloneData {
    pub bufList_: JSStructuredCloneData_BufferList,
    pub scope_: StructuredCloneScope,
    pub callbacks_: *const JSStructuredCloneCallbacks,
    pub closure_: *mut c_void,
    pub ownTransferables_: OwnTransferablePolicy,
    pub refsHeld_: SharedArrayRawBufferRefs,
}

JSStructuredCloneData represents structured clone data together with the information needed to read/write/transfer/free the records within it, in the form of a set of callbacks.

Fields
bufList_: JSStructuredCloneData_BufferListscope_: StructuredCloneScopecallbacks_: *const JSStructuredCloneCallbacksclosure_: *mut c_voidownTransferables_: OwnTransferablePolicyrefsHeld_: SharedArrayRawBufferRefs