Struct mozjs_sys::jsapi::JS::TransitiveCompileOptions

#[repr(C)]pub struct TransitiveCompileOptions {
    pub vtable_: *const TransitiveCompileOptions__bindgen_vtable,
    pub mutedErrors_: bool,
    pub forceFullParse_: bool,
    pub forceStrictMode_: bool,
    pub filename_: *const c_char,
    pub introducerFilename_: *const c_char,
    pub sourceMapURL_: *const u16,
    pub selfHostingMode: bool,
    pub extraWarningsOption: bool,
    pub werrorOption: bool,
    pub asmJSOption: AsmJSOption,
    pub throwOnAsmJSValidationFailureOption: bool,
    pub forceAsync: bool,
    pub discardSource: bool,
    pub sourceIsLazy: bool,
    pub allowHTMLComments: bool,
    pub hideScriptFromDebugger: bool,
    pub fieldsEnabledOption: bool,
    pub introductionType: *const c_char,
    pub introductionLineno: c_uint,
    pub introductionOffset: u32,
    pub hasIntroductionInfo: bool,
    pub instrumentationKinds: u32,
}

The common base class for the CompileOptions hierarchy.

Use this in code that needs to propagate compile options from one compilation unit to another.

Fields
vtable_: *const TransitiveCompileOptions__bindgen_vtablemutedErrors_: bool

The Web Platform allows scripts to be loaded from arbitrary cross-origin sources. This allows an attack by which a malicious website loads a sensitive file (say, a bank statement) cross-origin (using the user's cookies), and sniffs the generated syntax errors (via a window.onerror handler) for juicy morsels of its contents.

To counter this attack, HTML5 specifies that script errors should be sanitized ("muted") when the script is not same-origin with the global for which it is loaded. Callers should set this flag for cross-origin scripts, and it will be propagated appropriately to child scripts and passed back in JSErrorReports.

forceFullParse_: boolforceStrictMode_: boolfilename_: *const c_charintroducerFilename_: *const c_charsourceMapURL_: *const u16selfHostingMode: boolextraWarningsOption: boolwerrorOption: boolasmJSOption: AsmJSOptionthrowOnAsmJSValidationFailureOption: boolforceAsync: booldiscardSource: boolsourceIsLazy: boolallowHTMLComments: boolhideScriptFromDebugger: boolfieldsEnabledOption: boolintroductionType: *const c_char

|introductionType| is a statically allocated C string: one of "eval", "Function", or "GeneratorFunction".

introductionLineno: c_uintintroductionOffset: u32hasIntroductionInfo: boolinstrumentationKinds: u32