Type Definition mozjs_sys::jsapi::JS::Result

pub type Result = Result;

Result is intended to be the return type of JSAPI calls and internal functions that can run JS code or allocate memory from the JS GC heap. Such functions can:

succeed, possibly returning a value;

fail with a JS exception (out-of-memory falls in this category); or

fail because JS execution was terminated, which occurs when e.g. a user kills a script from the “slow script” UI. This is also how we unwind the stack when the debugger forces the current function to return. JS catch blocks can’t catch this kind of failure, and JS finally blocks don’t execute.