Module mozjs_sys::jsapi::mozilla

Modules

detail

ipc

span_details

tl

Structs

AlignedStorage2

AlignmentFinder

AlignmentFinder_Aligner

AutoCleanLinkedList

BaseTimeDuration

Instances of this class represent the length of an interval of time. Negative durations are allowed, meaning the end is before the start.

BaseTimeDurationPlatformUtils

Platform-specific implementation details of BaseTimeDuration.

BaseTimeDuration__SomethingVeryRandomHere

BitSet_Reference

BufferList

BufferList_IterImpl

BufferList_Segment

CStringHasher

CompactPair

CompactPair is the logical concatenation of an instance of A with an instance B. Space is conserved when possible. Neither A nor B may be a final class.

ContiguousEnumSize

ContiguousEnumValues

CorruptionCanary

This class is designed to cause crashes when various kinds of memory corruption are observed. For instance, let’s say we have a class C where we suspect out-of-bounds writes to some members. We can insert a member of type Poison near the members we suspect are being corrupted by out-of-bounds writes. Or perhaps we have a class K we suspect is subject to use-after-free violations, in which case it doesn’t particularly matter where in the class we add the member of type Poison.

CorruptionCanaryForStatics

A version of CorruptionCanary that is suitable as a member of objects that are statically allocated.

DebugOnly

DebugOnly contains a value of type T, but only in debug builds. In release builds, it does not contain a value. This helper is intended to be used with MOZ_ASSERT()-style macros, allowing one to write:

DefaultDelete

A default deletion policy using plain old operator delete.

DefaultHasher

EnumSet

EnumSet<T, U> is a set of values defined by an enumeration. It is implemented using a bit mask with the size of U for each value. It works both for enum and enum class types. EnumSet also works with U being a BitSet.

EnumSet_ConstIterator

EnumTypeFitsWithin

ErrorPropagationTag

A tag used to differentiate between GenericErrorResult created by the Err function (completely new error) and GenericErrorResult created by the Result::propagateErr function (propagated error). This can be used to track error propagation and eventually produce error stacks for logging/debugging purposes.

FallibleHashMethods

FloatingPoint

GenericErrorResult

A type that auto-converts to an error Result. This is like a Result without a success type. It’s the best return type for functions that always return an error–functions designed to build and populate error objects. It’s also useful in error-handling macros; see MOZ_TRY for an example.

HashCodeScrambler

A pseudorandom function mapping 32-bit integers to 32-bit integers.

HashCodeScrambler_SipHasher

HashMap

HashMapEntry

HashMap_MapHashPolicy

HashSet

HashSet_SetHashPolicy

LinkedList

LinkedListElement

LinkedList_Iterator

MallocAllocPolicy

MaxContiguousEnumValue

MaxEnumValue

Maybe

Maybe_SomeGuard

MinContiguousEnumValue

MovingNotNull

NeverAllocPolicy

NotNull

Nothing

Ok

Empty struct, indicating success for operations that have no return value. For example, if you declare another empty struct struct OutOfMemory {};, then Result<Ok, OutOfMemory> represents either success or OOM.

Opaque

Opaque is a replacement for integral T in cases where only comparisons must be supported, and it’s desirable to prevent accidental dependency on exact values.

OwningNonNull

PointerHasher

Range

RangedPtr

ReentrancyGuard

RefPtrTraits

Result

Result<V, E> represents the outcome of an operation that can either succeed or fail. It contains either a success value of type V or an error value of type E.

Span_storage_type

StaticLocalRefPtr

StaticRefPtr

TimeDurationValueCalculator

Perform arithmetic operations on the value of a BaseTimeDuration without doing strict checks on the range of values.

TimeStamp

TimeStampTests

UniquePtr

UniquePtr is a smart pointer that wholly owns a resource. Ownership may be transferred out of a UniquePtr through explicit action, but otherwise the resource is destroyed when the UniquePtr is destroyed.

Variant

mozilla::Variant

VariantType

Vector_CapacityAndReserved

Vector_ConstRange

Vector_Range

unused_t

Enums

LinkedListElement_NodeKind

MemoryOrdering

An enum of memory ordering possibilities for atomics.

NotNullTag

Vector__bindgen_ty_1

Constants

CorruptionCanaryForStatics_kCanarySet

Vector_InlineLength

Statics

VariantIndex_index

Type Aliases

Array_ElementType

Array_const_iterator

Array_const_reverse_iterator

Array_iterator

Array_reverse_iterator

AutoCleanLinkedList_ClientType

AutoCleanLinkedList_Traits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

CStringHasher_Key

CStringHasher_Lookup

CompactPair_Base

DefaultHasher_Lookup

EnumSet_serializedType

EnumSet_valueType

EnumeratedArray_ArrayType

EnumeratedArray_const_iterator

EnumeratedArray_const_reverse_iterator

EnumeratedArray_iterator

EnumeratedArray_reverse_iterator

FloatingPoint_Base

FloatingPoint_Bits

An unsigned integral type suitable for accessing the bitwise representation of T.

Generation

Opaque is a replacement for integral T in cases where only comparisons must be supported, and it’s desirable to prevent accidental dependency on exact values.

HashMapEntry_KeyType

HashMapEntry_ValueType

HashMap_AddPtr

HashMap_Entry

HashMap_Enum

HashMap_Impl

HashMap_Iterator

HashMap_Lookup

HashMap_MapHashPolicy_Base

HashMap_MapHashPolicy_KeyType

HashMap_ModIterator

HashMap_Ptr

HashMap_Range

HashMap_TableEntry

HashNumber

HashSet_AddPtr

HashSet_Entry

HashSet_Enum

HashSet_Impl

HashSet_Iterator

HashSet_Lookup

HashSet_ModIterator

HashSet_Ptr

HashSet_Range

HashSet_SetHashPolicy_Base

HashSet_SetHashPolicy_KeyType

InfinityBits_Traits

LinkedListElement_ClientType

LinkedListElement_ConstClientType

LinkedListElement_ConstRawType

LinkedListElement_RawType

LinkedListElement_Traits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

LinkedList_ClientType

LinkedList_ConstClientType

LinkedList_ConstElementType

LinkedList_ConstRawType

LinkedList_ElementType

LinkedList_Iterator_difference_type

LinkedList_Iterator_iterator_category

LinkedList_Iterator_pointer

LinkedList_Iterator_reference

LinkedList_Iterator_value_type

LinkedList_RawType

LinkedList_Traits

LinkedList supports refcounted elements using this adapter class. Clients using LinkedList<RefPtr> will get a data structure that holds a strong reference to T as long as T is in the list.

MallocSizeOf

Maybe_ValueType

PointerHasher_Lookup

Result_Impl

Result_err_type

Result_ok_type

Span_const_iterator

Span_const_reverse_iterator

Span_element_type

Span_index_type

Span_iterator

Span_pointer

Span_reference

Span_reverse_iterator

Span_value_type

SpecificNaNBits_Traits

TimeDuration

Specialization of BaseTimeDuration that uses TimeDurationValueCalculator for arithmetic on the mValue member.

TimeStampValue

UniquePtr_DeleterType

UniquePtr_ElementType

UniquePtr_Pointer

VariantType_Type

Variant_Impl

Variant_Tag

Vector_ElementType

Vector_Impl

Unions

AlignedStorage2_U

Utf8Unit

A code unit within a UTF-8 encoded string. (A code unit is the smallest unit within the Unicode encoding of a string. For UTF-8 this is an 8-bit number; for UTF-16 it would be a 16-bit number.)